// src/core/interfaces/importador.ts
export interface Importador {
  codigo_exportador: string;
  nif: string;
  local_embarque: string;
  codigo_regime: string;
  peso_bruto: string;
  forma_pagamento: string;
  numero_volumes: string;
  codigo_volumes: string;
  marcas_numeros: string;
  pais_precedentes: string;
  pais_destino: string;
}